# WIP: Move model data to class and overhaul