We're trying to use MVRSM to minimize a function with 340 decision variables, and we have observed a considerable performance hit when compared against other solvers like Hyperopt. This PR moves the model data away from a Python dictionary into a class, and refactors the surrogate model initialization, updating and minimization into said class. In addition, we overhauled some numpy-related code, replacing unnecessary vector transpositions to perform inner and outer products by `np.inner` and `np.outer`, respectively.

The exploration phase is currently WIP, since we have observed the results produced by the current code and this PR differ.This is caused because we no longer permute `next_x` during the first `rand_iters` iterations prior to generating a random `best_x` vector. If the current exploration code is preserved, then the results are the same. This was tested by copying the current code to a new file (say `MVRSM_current.py`) and modifying the `demo.py` file to look:
```python
# Line 17
import MVRSM_current
# [...[
# Line 338
	np.random.seed(0)
# [...]
# Line 361
	def run_MVRSM():
		np.random.seed(0)
		random.seed(0)
		solX_old, solY_old, model_old, logfile_old = MVRSM_current.MVRSM_minimize(obj_MVRSM, x0, lb, ub, num_int, max_evals, rand_evals)
		os.rename(logfile_old, os.path.join(folder, logfile_old))
		random.seed(0)
		np.random.seed(0)
		solX, solY, model, logfile = MVRSM.MVRSM_minimize(obj_MVRSM, x0, lb, ub, num_int, max_evals, rand_evals)
		os.rename(logfile, os.path.join(folder,logfile))
		print("Solution found: ")
		print(f"X = {solX}; {solX_old}")
		print(f"Y = {solY}; {solY_old}")
```
and then `diff`-ing the log files (e.g. with PyCharm).

We believe further optimizations are possible around the initialization code, but this is a negligible part of the runtime. It would also be interesting to vectorize the exploration phase.

The commented plotting and log file saving is preserved, although one could gate this functionality under flags passed to the `MVRSM_minimize` function. This would prevent opening and writing log files to disk when they aren't needed.

Unrelated, but the original code was straightforward to read as it follows the paper quite closely. Also, the dependency-related stuff was easy to setup when compared with othe research projects (heh!). Thank you :)