@hsanzg 2022-10-24
50b4c244c92a9c853b8dfcbe662b62a7fc915323 solves a bug where we were ignoring the output of the safeguard `np.clip` calls, which are now performed in place.