@hsanzg 2022-10-24
Some preliminary results: the iteration-vs-objective plots produced by the current code and this PR are extremely close, suggesting that the current behavior is maintained, bar the random exploration. 30 trials of 53-variable Rosenbrock problem with 100 iterations give
Current:
![img](https://user-images.githubusercontent.com/80487270/197495654-ff271f5a-50f6-4967-90ea-d96949963a92.png)

This PR:
![img_1](https://user-images.githubusercontent.com/80487270/197495701-8d0a9159-afc5-410f-884c-1926c75fd962.png)

The computation time is reduced considerably. In our problem with 340 discrete decision variables we have observed the median iteration times drop from 1.95s to 0.21s. The update times remain more or less the same, but the minimization step is reduced from 1.81s to 0.09s, due to the optimized objective and Jacobian expressions.