@hsanzg 2022-10-24
I believe I've fixed a subtle bug in 2070894bacd71bfe2ea2fd6a725813600033b313, but I'm not 100% sure about it. When initializing the model, the weights `c_i` corresponding to discrete variables are set to 1. Currently, this is done with
```python
for i in range(1, num_discr_basisfunctions):
    c[i] = 1
```
where `num_discr_basisfunctions = len(B) - 1` (# of basis functions related to discrete variables minus the constant basis function independent of x). The `stop` argument of Python's `range` function is exclusive, thus this sets all the weights except the last one (at index `len(B) - 1`) to one. `c_{len(B) - 1}` corresponds to the last basis function dependent on two subsequent discrete variables, hence I believe it should also be set to 1. Is this correct?